compilation errors resolved. msg_ & ik_mes alive
changed matlab (a^2) to pow(a,2) etc.
abs()->fabs().
The specific move task needs an extended message... I will extend the
used msg_message_s to resemble the current ik_message_s. The current
msg_message_s is a subset of ik_message_s so the rest of the code should
work afterwards...