Servo speed set individually ->  =time start stop
All servo speeds are calculated individually within the ik function
using the goal pwm value and the current pwm value. This is done by
taking the longest move, calculating its time needed when observing the
standart speed of 12500 as set in movement.h(hardcoded to that value
atm...) and calculating the speed for each other servo using its
distance over the maximum time.
There are only 11 bits available in the msgData - therefore 12500/6.11
<=2048(2^11). This needs to be decrypted on the other side by
multiplying with 6.11 again. We obviously loose some resolution...